requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0
Werkzeug==3.0.1
SQLAlchemy==2.0.27
alembic==1.13.1
python-dateutil==2.8.2